add current user depends for endpoint
add error handler for RTK request

+++ backend/app/routes/history.py
@@ -1,12 +1,13 @@
 import logging
 
-from fastapi import APIRouter, Body, HTTPException, status
+from fastapi import APIRouter, Body, HTTPException, status, Depends
 from fastapi.encoders import jsonable_encoder
 from uvicorn.server import logger
 
+from authorisation.auth import get_current_active_user
 from utils.db import add_history_item, retrieve_user, retrieve_user_by_name, retrieve_history_item_by_username, \
     delete_history_by_username_and_query
-from utils.schema import HistorySchema, ResponseModel, ErrorResponseModel, ResponseHistoryModel, HistoryBaseSchema
+from utils.schema import HistorySchema, ResponseModel, ErrorResponseModel, ResponseHistoryModel, HistoryBaseSchema, User
 
 router = APIRouter()
 
@@ -29,7 +30,8 @@ async def get_history_by_username(username: str):
     if not user:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail='Не найден пользователь. Проверьте правильность запроса.'
+            detail='Не найден пользователь. Проверьте правильность запроса.',
+            headers = {'WWW-Authenticate': 'Bearer'}
         )
     history_items = await retrieve_history_item_by_username(username)
     if history_items:
@@ -37,6 +39,14 @@ async def get_history_by_username(username: str):
     return ResponseHistoryModel([], 'Записи в истории не найдены')
 
 @router.post("/delete")
-async def delete_history(params: HistoryBaseSchema = Body(...)):
-    await delete_history_by_username_and_query(username=params.username, query=params.query)
-    return ResponseModel({'details': 'Запрос из истории удален'}, 'Ok')
+async def delete_history(params: HistoryBaseSchema = Body(...), current_user: User = Depends(get_current_active_user)):
+    if current_user['username'] == params.username:
+        await delete_history_by_username_and_query(username=params.username, query=params.query)
+        return ResponseModel({'details': 'Запрос из истории удален'}, 'Ok')
+
+    raise HTTPException(
+        status_code=status.HTTP_400_BAD_REQUEST,
+        detail='Авторизованный пользователь от которого происходи запрос не '
+               'совпадает пользователем запросы которого вы хотите удалить.',
+        headers={'WWW-Authenticate': 'Bearer'}
+    )
